Add files via upload
I attach my final project for Quantitative Macroeocnomics.

The title is:
What is the labour response to a technology shock and what is the European Central Bank’s response in the New Keynesian Model?

In this study I started to develop a first approach in the research field I am interested in. The aim of the study was to explain how the technology shock affects the hours worked by individuals, and also, the possible response of Central Bank. Regarding the methodology, Taylor approach was used and the code was run in MATLAB.